feat: use new appflowy-editor-plugins package
