+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_page.dart
@@ -1,6 +1,9 @@
+import 'dart:ui' as ui;
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
+import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy/plugins/document/application/document_bloc.dart';
 import 'package:appflowy/plugins/document/presentation/editor_configuration.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/align_toolbar_item/custom_text_align_command.dart';
@@ -21,14 +24,32 @@ import 'package:appflowy/workspace/application/settings/shortcuts/settings_short
 import 'package:appflowy/workspace/application/view_info/view_info_bloc.dart';
 import 'package:appflowy/workspace/presentation/settings/widgets/emoji_picker/emoji_picker.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
+import 'package:appflowy_editor_plugins/appflowy_editor_plugins.dart';
 import 'package:collection/collection.dart';
+import 'package:easy_localization/easy_localization.dart';
 import 'package:flowy_infra/theme_extension.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+final codeBlockLocalization = CodeBlockLocalizations(
+  codeBlockNewParagraph:
+      LocaleKeys.settings_shortcuts_commands_codeBlockNewParagraph.tr(),
+  codeBlockAddTwoSpaces:
+      LocaleKeys.settings_shortcuts_commands_codeBlockAddTwoSpaces.tr(),
+  codeBlockDeleteTwoSpaces:
+      LocaleKeys.settings_shortcuts_commands_codeBlockDeleteTwoSpaces.tr(),
+  codeBlockSelectAll:
+      LocaleKeys.settings_shortcuts_commands_codeBlockSelectAll.tr(),
+  codeBlockPasteText:
+      LocaleKeys.settings_shortcuts_commands_codeBlockPasteText.tr(),
+);
+
+final localizedCodeBlockCommands =
+    codeBlockCommands(localizations: codeBlockLocalization);
+
 final List<CommandShortcutEvent> commandShortcutEvents = [
   toggleToggleListCommand,
-  ...codeBlockCommands,
+  ...localizedCodeBlockCommands,
   customCopyCommand,
   customPasteCommand,
   customCutCommand,
@@ -90,7 +111,7 @@ class _AppFlowyEditorPageState extends State<AppFlowyEditorPage> {
 
   late final List<CommandShortcutEvent> commandShortcutEvents = [
     toggleToggleListCommand,
-    ...codeBlockCommands,
+    ...localizedCodeBlockCommands,
     customCopyCommand,
     customPasteCommand,
     customCutCommand,
@@ -264,7 +285,7 @@ class _AppFlowyEditorPageState extends State<AppFlowyEditorPage> {
     final isRTL =
         context.read<AppearanceSettingsCubit>().state.layoutDirection ==
             LayoutDirection.rtlLayout;
-    final textDirection = isRTL ? TextDirection.rtl : TextDirection.ltr;
+    final textDirection = isRTL ? ui.TextDirection.rtl : ui.TextDirection.ltr;
 
     _setRTLToolbarItems(
       context.read<AppearanceSettingsCubit>().state.enableRtlToolbarItems,
