--- frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/code_block/code_block_themes.dart
@@ -1,116 +1,0 @@
-import 'package:flutter/painting.dart';
-
-const lightThemeInCodeblock = {
-  'root': TextStyle(
-    backgroundColor: Color(0xfffbf1c7),
-    color: Color(0xff3c3836),
-  ),
-  'subst': TextStyle(color: Color(0xff3c3836)),
-  'deletion': TextStyle(color: Color(0xff9d0006)),
-  'formula': TextStyle(color: Color(0xff9d0006)),
-  'keyword': TextStyle(color: Color(0xff9d0006)),
-  'link': TextStyle(color: Color(0xff9d0006)),
-  'selector-tag': TextStyle(color: Color(0xff9d0006)),
-  'built_in': TextStyle(color: Color(0xff076678)),
-  'emphasis': TextStyle(
-    color: Color(0xff076678),
-    fontStyle: FontStyle.italic,
-  ),
-  'name': TextStyle(color: Color(0xff076678)),
-  'quote': TextStyle(color: Color(0xff076678)),
-  'strong': TextStyle(
-    color: Color(0xff076678),
-    fontWeight: FontWeight.bold,
-  ),
-  'title': TextStyle(color: Color(0xff076678)),
-  'variable': TextStyle(color: Color(0xff076678)),
-  'attr': TextStyle(color: Color(0xffb57614)),
-  'params': TextStyle(color: Color(0xffb57614)),
-  'template-tag': TextStyle(color: Color(0xffb57614)),
-  'type': TextStyle(color: Color(0xffb57614)),
-  'builtin-name': TextStyle(color: Color(0xff8f3f71)),
-  'doctag': TextStyle(color: Color(0xff8f3f71)),
-  'literal': TextStyle(color: Color(0xff8f3f71)),
-  'number': TextStyle(color: Color(0xff8f3f71)),
-  'code': TextStyle(color: Color(0xffaf3a03)),
-  'meta': TextStyle(color: Color(0xffaf3a03)),
-  'regexp': TextStyle(color: Color(0xffaf3a03)),
-  'selector-id': TextStyle(color: Color(0xffaf3a03)),
-  'template-variable': TextStyle(color: Color(0xffaf3a03)),
-  'addition': TextStyle(color: Color(0xff79740e)),
-  'meta-string': TextStyle(color: Color(0xff79740e)),
-  'section': TextStyle(
-    color: Color(0xff79740e),
-    fontWeight: FontWeight.bold,
-  ),
-  'selector-attr': TextStyle(color: Color(0xff79740e)),
-  'selector-class': TextStyle(color: Color(0xff79740e)),
-  'string': TextStyle(color: Color(0xff79740e)),
-  'symbol': TextStyle(color: Color(0xff79740e)),
-  'attribute': TextStyle(color: Color(0xff427b58)),
-  'bullet': TextStyle(color: Color(0xff427b58)),
-  'class': TextStyle(color: Color(0xff427b58)),
-  'function': TextStyle(color: Color(0xff427b58)),
-  'meta-keyword': TextStyle(color: Color(0xff427b58)),
-  'selector-pseudo': TextStyle(color: Color(0xff427b58)),
-  'tag': TextStyle(
-    color: Color(0xff427b58),
-    fontWeight: FontWeight.bold,
-  ),
-  'comment': TextStyle(
-    color: Color(0xff928374),
-    fontStyle: FontStyle.italic,
-  ),
-  'link_label': TextStyle(color: Color(0xff8f3f71)),
-};
-
-const darkThemeInCodeBlock = {
-  'root': TextStyle(
-    backgroundColor: Color(0xff000000),
-    color: Color(0xfff8f8f8),
-  ),
-  'comment': TextStyle(
-    color: Color(0xffaeaeae),
-    fontStyle: FontStyle.italic,
-  ),
-  'quote': TextStyle(
-    color: Color(0xffaeaeae),
-    fontStyle: FontStyle.italic,
-  ),
-  'keyword': TextStyle(color: Color(0xffe28964)),
-  'selector-tag': TextStyle(color: Color(0xffe28964)),
-  'type': TextStyle(color: Color(0xffe28964)),
-  'string': TextStyle(color: Color(0xff65b042)),
-  'subst': TextStyle(color: Color(0xffdaefa3)),
-  'regexp': TextStyle(color: Color(0xffe9c062)),
-  'link': TextStyle(color: Color(0xffe9c062)),
-  'title': TextStyle(color: Color(0xff89bdff)),
-  'section': TextStyle(color: Color(0xff89bdff)),
-  'tag': TextStyle(color: Color(0xff89bdff)),
-  'name': TextStyle(color: Color(0xff89bdff)),
-  'symbol': TextStyle(color: Color(0xff3387cc)),
-  'bullet': TextStyle(color: Color(0xff3387cc)),
-  'number': TextStyle(color: Color(0xff3387cc)),
-  'params': TextStyle(color: Color(0xff3e87e3)),
-  'variable': TextStyle(color: Color(0xff3e87e3)),
-  'template-variable': TextStyle(color: Color(0xff3e87e3)),
-  'attribute': TextStyle(color: Color(0xffcda869)),
-  'meta': TextStyle(color: Color(0xff8996a8)),
-  'formula': TextStyle(
-    backgroundColor: Color(0xff0e2231),
-    color: Color(0xfff8f8f8),
-    fontStyle: FontStyle.italic,
-  ),
-  'addition': TextStyle(
-    backgroundColor: Color(0xff253b22),
-    color: Color(0xfff8f8f8),
-  ),
-  'deletion': TextStyle(
-    backgroundColor: Color(0xff420e09),
-    color: Color(0xfff8f8f8),
-  ),
-  'selector-class': TextStyle(color: Color(0xff9b703f)),
-  'selector-id': TextStyle(color: Color(0xff8b98ab)),
-  'emphasis': TextStyle(fontStyle: FontStyle.italic),
-  'strong': TextStyle(fontWeight: FontWeight.bold),
-};
