+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/plugins.dart
@@ -3,8 +3,7 @@ export 'actions/option_action.dart';
 export 'align_toolbar_item/align_toolbar_item.dart';
 export 'base/toolbar_extension.dart';
 export 'callout/callout_block_component.dart';
-export 'code_block/code_block_component.dart';
-export 'code_block/code_block_shortcut_event.dart';
+export 'code_block/code_block_language_selector.dart';
 export 'context_menu/custom_context_menu.dart';
 export 'copy_and_paste/custom_copy_command.dart';
 export 'copy_and_paste/custom_cut_command.dart';
